face_recognition/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
matplotlib>=3.4.2
scikit-learn>=0.24.2
tqdm>=4.61.1
Pillow>=8.2.0
timm>=0.5.4 